server/pyproject.toml
[tool.poetry]
name = "pixelsphere"
version = "0.1.0"
description = ""
authors = ["tracebn"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "0.99.1"
pydantic = "1.10.11"
uvicorn = "^0.26.0"
alembic = "^1.13.1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
python-multipart = "^0.0.6"
pytz = "^2023.3.post1"
fastapi-jwt-auth = {extras = ["asymmetric"], version = "^0.5.0"}
passlib = "^1.7.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
